+++ src/extension/components/NewAccountItem/types/IProps.ts
@@ -1,0 +1,8 @@
+interface IProps {
+  address: string;
+  name?: string;
+  subTextColor?: string;
+  textColor?: string;
+}
+
+export default IProps;