+++ src/extension/translations/en.ts
@@ -427,6 +427,8 @@ const translation: IResourceLanguage = {
     selectAssets: 'Select Assets',
     selectANetwork: 'Select A Network',
     selectAnOption: 'Select An Option',
+    selectColor: 'Select Color',
+    selectIcon: 'Select Icon',
     selectReceiverAccount: 'Select Receiver Account',
     selectSenderAccount: 'Select Sender Account',
     sendAsset: 'Send {{asset}}',