+++ src/extension/types/accounts/index.ts
@@ -5,3 +5,5 @@ export type { default as IAccountWithExtendedProps } from './IAccountWithExtende
 export type { default as IActiveAccountDetails } from './IActiveAccountDetails';
 export type { default as IInitializeAccountOptions } from './IInitializeAccountOptions';
 export type { default as INewAccount } from './INewAccount';
+export type { default as TAccountColors } from './TAccountColors';
+export type { default as TAccountIcons } from './TAccountIcons';